Splitting centralized_telescope in files in new folder
