+++ src/centralized_telescope/types.rs
@@ -1,0 +1,3 @@
+pub(crate) const DIGEST_SIZE: usize = 32;
+
+pub(super) type Hash = [u8; DIGEST_SIZE];
