+++ src/lib.rs
@@ -1,8 +1,5 @@
 //! An implementation of Approximate Lower Bound Arguments
 //! (ALBA, <https://eprint.iacr.org/2023/1655.pdf>).
-#[cfg(test)]
-mod test_utils;
-
 mod utils;
 
 pub mod centralized_telescope;
