+++ src/utils/types.rs
@@ -1,0 +1,3 @@
+pub(crate) const DATA_LENGTH: usize = 32;
+
+pub(crate) type Element = [u8; DATA_LENGTH];
